fix(lint): move AnyJsGetter and AnyJsSetter to biome_js_syntax
